Update `cache-handler-redis` example
