+++ examples/cache-handler-redis/package.json
@@ -13,7 +13,7 @@
   "devDependencies": {
     "@neshca/cache-handler": "latest",
     "@neshca/json-replacer-reviver": "latest",
-    "@types/node": "^20.9.0",
+    "@types/node": "^20.9.1",
     "@types/react": "^18.2.37",
     "@types/react-dom": "^18.2.15",
     "redis": "latest",
